pyproject.toml
[build-system]
requires = ['flit_core >=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'aiida-sdb'
dynamic = ['version']
description = 'Package for the SDB project.'
authors = [{name = 'Sebastiaan P. Huber', email = 'mail@sphuber.net'}]
readme = 'README.md'
license = {file = 'LICENSE.txt'}
classifiers = [
    'Development Status :: 4 - Development',
    'Framework :: AiiDA',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]
keywords = ['aiida', 'workflows']
requires-python = '>=3.8'
dependencies = [
    'aiida-core[atomic_tools]~=2.3.0',
    'aiida-codtools~=3.0',
    'click~=8.0',
    'numpy',
    'scipy',
    'typer[all]~=0.9'
]

[project.urls]
Home = 'https://github.com/sphuber/aiida-sdb'
Source = 'https://github.com/sphuber/aiida-sdb'

[project.optional-dependencies]
dev = [
    'pre-commit~=2.17',
    'pylint~=2.12.2',
    'pylint-aiida~=0.1',
    'toml'
]

[project.scripts]
aiida-sdb = 'aiida_sdb.cli:cmd_root'
mc3d = 'aiida_sdb.cli.main:app'

[tool.flit.module]
name = 'aiida_sdb'

[tool.isort]
profile = "black"
